aider: Added functionality to detect changes in `currentCCValue` and transmit a MIDI CC message with the interpolated value of the new `currentCCValue`.
# Aider chat conversation:

USER: we should detect when currentCCValue has changed and transmit a midi cc message on midiChannelOut and ccNumber of the interpolated value of the new currentCCValue.

ASSISTANT: To implement this, we need to add a new variable `previousCCValue` to keep track of the last `currentCCValue`. Then, in the `@block` section, we can compare `currentCCValue` with `previousCCValue`. If they are not equal, we calculate the interpolated value and transmit a MIDI CC message. After that, we update `previousCCValue` with the new `currentCCValue`.

Here are the changes:

```js
ccinterpolate
<<<<<<< HEAD
@init
function interpolate(cc, startValue, endValue) (
  startValue + (cc * (endValue - startValue)) / 127;
);

@slider
currentCCValue = slider1;
startValue = slider2;
endValue = slider3;
midiChannelOut = slider4;
ccNumber = slider5;

@block
// Per-block processing code here
=======
@init
function interpolate(cc, startValue, endValue) (
  startValue + (cc * (endValue - startValue)) / 127;
);
var previousCCValue = undefined;

@slider
currentCCValue = slider1;
startValue = slider2;
endValue = slider3;
midiChannelOut = slider4;
ccNumber = slider5;

@block
if (currentCCValue != previousCCValue) {
  var interpolatedValue = interpolate(currentCCValue, startValue, endValue);
  midisend(midiChannelOut, ccNumber, interpolatedValue);
  previousCCValue = currentCCValue;
}
>>>>>>> updated
```

Please note that `midisend` is a placeholder function. Replace it with the actual function to send a MIDI CC message. The function should take three parameters: the MIDI channel, the CC number, and the value.